requirements.txt
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.2
gunicorn==19.9.0
Markdown==3.1.1
pytz==2019.2
sqlparse==0.3.0
